DaisyTrade/daisy_bot/requirements.txt
python-binance
pandas
numpy
TA-Lib
